themes/wwtguide: update to latest `master`
Note that we don't need to do this to fix the Azure CI, since it pulls
the Azure Pipelines script from the separate repo without caring about
what version we have checked out in our submodule. But I need to rerun
the CI anyway, so I might as well update.